+++ src/stackexchange/mod.rs
@@ -5,5 +5,5 @@ mod search;
 pub mod scraper;
 
 pub use api::{Answer, Id, Question};
-pub use local_storage::LocalStorage;
+pub use local_storage::{LocalStorage, SiteMap};
 pub use search::Search;